requirements.txt
beautifulsoup4==4.12.2
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
constantly==15.1.0
cssselect==1.2.0
cymem==2.0.7
dotmap==1.3.30
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.11.0
Flask==2.2.3
hjson==3.1.0
huggingface-hub==0.13.4
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.2
mpmath==1.2.1
murmurhash==1.0.9
networkx==3.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.1
packaging==23.0
pathy==0.10.1
Pillow==9.5.0
plac==1.3.5
preshed==3.0.8
pyasn1==0.4.8
pydantic==1.10.7
PyDispatcher==2.0.7
python-dateutil==2.8.2
pywin32==306
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
sentencepiece==0.1.97
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.11.1
thinc==8.1.9
tinysegmenter==0.3
tldextract==3.4.0
tokenizers==0.13.3
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tqdm==4.65.0
transformers==4.27.4
twisted-iocpsupport==1.0.3
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.15
wasabi==1.1.1
Werkzeug==2.2.3
